fix(hardening): EPS harmonization, mypy gate, Zhou+Prony guards, anchor tests
Four hardening candidates plus a debug pass:

1. EPS harmonization (nonnormality.petermann_factors): replace hard-coded
   1.0e-300 divide-by-zero floor with a named, shared _consts.EPS_DIV. Kept
   the value (underflow floor, NOT a physics tolerance) so no anchor changes;
   verified canonical fixtures have |<l,r>|^2 ~ O(1), far from any threshold.

2. mypy CI gate: removed continue-on-error in ci.yml; fixed all 18 prior
   type-blind findings (annotated numpy return locals, is_dataclass instance
   narrowing in io/export). No behaviour change. mypy now exits 0.

3. Zhou guard (_zhou): adopt the canonical EPS_DIV/EPS_GAP constants for the
   Petermann-style division floor and zero-eigenvalue check. New closed-form
   anchor test (single-qubit pure dephasing: gap=1, K=1 -> both bounds =
   log(1/eps)) + a defective-mode guard test.

4. Prony fallback (fitting.prony): wrap lstsq/np.roots in try/except, emit a
   RuntimeWarning and fall back to a safe default seed (positive amplitude)
   on near-singular/non-finite Hankel data. fails-before: the pre-guard code
   raised LinAlgError on all-NaN/inf input (proven by stashing prony.py).

Debug pass: property-probed fitting/ + diagnostics/ (empty/single/NaN/inf
inputs). Prony was the one clear, provable bug. choose_model({})->'M0' and
estimate_neff_geyer(NaN)->n are documented/defensible, left as observations.

Local gates: ruff 0, mypy 0, pytest 162 passed / 4 skipped, coverage 87.95%
(>=80). Anchor regressions 17 passed / 2 skipped, 0 anchor values changed.

Backup-First context: no destructive git ops (ref Working-agreements §2,
2026-05-16 data-loss incident) -- branch-create + commit + push only.

+++ src/liouscope/fitting/aicc.py
@@ -40,7 +40,7 @@ def gaussian_log_likelihood(
         sigma_sq = max(rss / n, 1.0e-30)
     else:
         sigma_sq = sigma * sigma
-    return -0.5 * n * (np.log(2.0 * np.pi * sigma_sq) + rss / (n * sigma_sq))
+    return float(-0.5 * n * (np.log(2.0 * np.pi * sigma_sq) + rss / (n * sigma_sq)))
 
 
 def choose_model(aiccs: Mapping[str, float]) -> str:
